pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dynamoui"
version = "1.0.0"
description = "Adaptive Data Interface Framework — LLM-powered UI generation for backend data models"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy>=2.0.29",
    "asyncpg>=0.29.0",
    "aiofiles>=23.2.1",
    "pyyaml>=6.0.1",
    "rapidfuzz>=3.9.0",
    "structlog>=24.1.0",
    "click>=8.1.7",
    "anthropic>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "alembic>=1.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.3.0",
]

[project.scripts]
dynamoui = "backend.skill_registry.cli.validate:cli"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["backend"]
omit = ["*/tests/*", "*/__init__.py"]
